controllers/user/update: Move `regenerate_token_and_send()` to a dedicated module
