+++ src/controllers/user.rs
@@ -1,6 +1,8 @@
 pub mod me;
 pub mod other;
+mod resend;
 pub mod session;
 pub mod update;
 
+pub use resend::regenerate_token_and_send;
 pub use update::update_user;
