+++ src/controllers/user/update.rs
@@ -1,7 +1,7 @@
 use crate::app::AppState;
 use crate::auth::AuthCheck;
 use crate::controllers::helpers::ok_true;
-use crate::models::{Email, NewEmail};
+use crate::models::NewEmail;
 use crate::schema::{emails, users};
 use crate::tasks::spawn_blocking;
 use crate::util::diesel::prelude::*;
@@ -9,7 +9,6 @@ use crate::util::errors::{bad_request, server_error, AppResult};
 use axum::extract::Path;
 use axum::response::Response;
 use axum::Json;
-use diesel::dsl::sql;
 use diesel_async::async_connection_wrapper::AsyncConnectionWrapper;
 use http::request::Parts;
 use lettre::Address;
@@ -108,47 +107,6 @@ pub async fn update_user(
 
             let _ = state.emails.send(user_email, email);
         }
-
-        ok_true()
-    })
-    .await
-}
-
-/// Handles `PUT /user/:user_id/resend` route
-pub async fn regenerate_token_and_send(
-    state: AppState,
-    Path(param_user_id): Path<i32>,
-    req: Parts,
-) -> AppResult<Response> {
-    let mut conn = state.db_write().await?;
-    let auth = AuthCheck::default().check(&req, &mut conn).await?;
-    spawn_blocking(move || {
-        use diesel::RunQueryDsl;
-
-        let conn: &mut AsyncConnectionWrapper<_> = &mut conn.into();
-
-        let user = auth.user();
-
-        // need to check if current user matches user to be updated
-        if user.id != param_user_id {
-            return Err(bad_request("current user does not match requested user"));
-        }
-
-        conn.transaction(|conn| -> AppResult<_> {
-            let email: Email = diesel::update(Email::belonging_to(user))
-                .set(emails::token.eq(sql("DEFAULT")))
-                .get_result(conn)
-                .optional()?
-                .ok_or_else(|| bad_request("Email could not be found"))?;
-
-            let email1 = UserConfirmEmail {
-                user_name: &user.gh_login,
-                domain: &state.emails.domain,
-                token: email.token,
-            };
-
-            state.emails.send(&email.email, email1).map_err(Into::into)
-        })?;
 
         ok_true()
     })
