+++ src/router.rs
@@ -133,7 +133,7 @@ pub fn build_axum_router(state: AppState) -> Router<()> {
         )
         .route(
             "/api/v1/users/:user_id/resend",
-            put(user::update::regenerate_token_and_send),
+            put(user::regenerate_token_and_send),
         )
         .route(
             "/api/v1/site_metadata",
